+++ cmds/config-testnet-supernode.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 set -e
-symbol-bootstrap config -p testnet -a dual -t target/testnet-supernode -c test/supernode.yml --noPassword $1 $2 $3
+symbol-bootstrap config -p testnet -a dual -t target/testnet-supernode -c test/supernode.yml --password 1111  $1 $2 $3