+++ cmds/start-testnet-supernode-demo.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+symbol-bootstrap start -p testnet -a demo -t target/testnet-supernode -c test/supernode.yml  $1 $2 $3