+++ src/service/BootstrapUtils.ts
@@ -32,7 +32,7 @@ import * as Handlebars from 'handlebars';
 import { get } from 'https';
 import * as _ from 'lodash';
 import { platform, totalmem } from 'os';
-import { basename, join, resolve } from 'path';
+import { basename, dirname, join, resolve } from 'path';
 import { Convert, DtoMapping, NetworkType } from 'symbol-sdk';
 import * as util from 'util';
 import { LogType } from '../logger';
@@ -373,6 +373,21 @@ export class BootstrapUtils {
         );
     }
 
+    public static async chmodRecursive(path: string, mode: string | number): Promise<void> {
+        // Loop through all the files in the config folder
+        const stat = await fsPromises.stat(path);
+        if (stat.isFile()) {
+            await fsPromises.chmod(path, mode);
+        } else if (stat.isDirectory()) {
+            const files = await fsPromises.readdir(path);
+            await Promise.all(
+                files.map(async (file: string) => {
+                    await this.chmodRecursive(join(path, file), mode);
+                }),
+            );
+        }
+    }
+
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
     public static runTemplate(template: string, templateContext: any): string {
         const compiledTemplate = Handlebars.compile(template);
@@ -381,6 +396,13 @@ export class BootstrapUtils {
 
     public static async mkdir(path: string): Promise<void> {
         await fsPromises.mkdir(path, { recursive: true });
+    }
+
+    public static async mkdirParentFolder(fileName: string): Promise<void> {
+        const parentFolder = dirname(fileName);
+        if (parentFolder) {
+            await BootstrapUtils.mkdir(parentFolder);
+        }
     }
 
     // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
@@ -438,6 +460,7 @@ export class BootstrapUtils {
     }
 
     public static async writeTextFile(path: string, text: string): Promise<void> {
+        await BootstrapUtils.mkdirParentFolder(path);
         await fsPromises.writeFile(path, text, 'utf8');
     }