+++ test/service/ConfigService.test.ts
@@ -16,7 +16,7 @@
 
 import { expect } from '@oclif/test';
 import 'mocha';
-import { ConfigService, Preset } from '../../src/service';
+import { ConfigService, CryptoUtils, Preset } from '../../src/service';
 
 describe('ConfigService', () => {
     it('ConfigService default run with optin_preset.yml', async () => {
@@ -58,12 +58,23 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService bootstrap default', async () => {
-        await new ConfigService('.', {
+        const configResult = await new ConfigService('.', {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/bootstrap',
             preset: Preset.bootstrap,
         }).run();
+
+        const configResultUpgrade = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            upgrade: true,
+            target: 'target/tests/bootstrap',
+            preset: Preset.bootstrap,
+        }).run();
+
+        expect(CryptoUtils.removePrivateKeys(configResultUpgrade.presetData)).deep.eq(
+            CryptoUtils.removePrivateKeys(configResult.presetData),
+        );
     });
 
     it('ConfigService bootstrap repeat', async () => {
@@ -89,5 +100,65 @@ describe('ConfigService', () => {
         assertRepeatedService(7, configResult.presetData.nodes);
         assertRepeatedService(4, configResult.presetData.gateways);
         assertRepeatedService(4, configResult.presetData.databases);
+    });
+
+    it('ConfigService mainnet supernode assembly after upgrade', async () => {
+        const configResultInitial = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            reset: true,
+            target: 'target/tests/ConfigService.mainnet.supernode',
+            preset: Preset.mainnet,
+            customPreset: 'test/unit-test-profiles/supernode.yml',
+            assembly: 'dual',
+        }).run();
+        const configResultUpgrade = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            upgrade: true,
+            target: 'target/tests/ConfigService.mainnet.supernode',
+            preset: Preset.mainnet,
+            assembly: 'dual',
+        }).run();
+        expect(CryptoUtils.removePrivateKeys(configResultUpgrade.presetData)).deep.eq(
+            CryptoUtils.removePrivateKeys(configResultInitial.presetData),
+        );
+        const configResultUpgradeSecond = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            upgrade: true,
+            target: 'target/tests/ConfigService.mainnet.supernode',
+        }).run();
+
+        expect(CryptoUtils.removePrivateKeys(configResultUpgradeSecond.presetData)).deep.eq(
+            CryptoUtils.removePrivateKeys(configResultInitial.presetData),
+        );
+    });
+
+    it('ConfigService testnet supernode assembly after upgrade', async () => {
+        const configResultInitial = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            reset: true,
+            target: 'target/tests/ConfigService.testnet.supernode',
+            preset: Preset.testnet,
+            customPreset: 'test/unit-test-profiles/supernode.yml',
+            assembly: 'dual',
+        }).run();
+        const configResultUpgrade = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            upgrade: true,
+            target: 'target/tests/ConfigService.testnet.supernode',
+            preset: Preset.testnet,
+            assembly: 'dual',
+        }).run();
+        expect(CryptoUtils.removePrivateKeys(configResultUpgrade.presetData)).deep.eq(
+            CryptoUtils.removePrivateKeys(configResultInitial.presetData),
+        );
+        const configResultUpgradeSecond = await new ConfigService('.', {
+            ...ConfigService.defaultParams,
+            upgrade: true,
+            target: 'target/tests/ConfigService.testnet.supernode',
+        }).run();
+
+        expect(CryptoUtils.removePrivateKeys(configResultUpgradeSecond.presetData)).deep.eq(
+            CryptoUtils.removePrivateKeys(configResultInitial.presetData),
+        );
     });
 });