+++ CHANGELOG.md
@@ -4,7 +4,45 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.3] - NEXT
+## [1.0.6] - Next
+
+**Milestone**: Mainnet(1.0.0.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.0.6  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   Added `MonitorOnly` reward program.
+
+## [1.0.5] - May-3-2021
+
+**Milestone**: Mainnet(1.0.0.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.0.5  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   Re-enabled node reward program. Upgraded TLS communication.
+-   Fixed `failed to load: /docker-entrypoint-initdb.d/mongoDbPrepare.js` when running with root user.
+-   Renamed `enrol` to `enroll` for the Rewards Program.
+
+## [1.0.4] - Apr-13-2021
+
+**Milestone**: Mainnet(1.0.0.0)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v1.0.4  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   New `testnet.symboldev.network` testnet!
+-   Added `verify` command.
+-   Fixed host override when no custom preset is provided in mainnet.
+-   Fixed case issue validating keys when creating certificates.
+-   Updated Wallet to latest 1.0.1 release.
+-   Node properties sinkType: Async and enableSingleThreadPool: false by default in peer nodes too.
+-   Dropped NodeJS 10 support. Added Node LTS and Stable Support. 
+
+## [1.0.3] - Mar-31-2021
 
 **Milestone**: Mainnet(1.0.0.0)
 
@@ -14,6 +52,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 
 -   Native and offline ready voting key file generation.
 -   Improved Custom Preset Object types for symbol bootstrap lib integration.
+-   TransactionSelectionStrategy's new default value is `oldest`.
 
 ## [1.0.2] - Mar-24-2021